+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:device_preview/device_preview.dart';
+import 'package:movie_curation/ui/screens/root/root_screen_tablet.dart';
 import 'utilities/index.dart';
 
 // Device Preview 셋업 (반응형 확인)
@@ -10,7 +11,7 @@ import 'utilities/index.dart';
 //     );
 
 void main() {
-  FRouter.setupRouter();
+  // FRouter.setupRouter();
   runApp(const MyApp());
 }
 
@@ -19,10 +20,6 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // SystemChrome.setPreferredOrientations([
-    //   DeviceOrientation.portraitUp,
-    //   DeviceOrientation.portraitDown,
-    // ]);
     return Sizer(builder: (context, orientation, deviceType) {
       return GetMaterialApp(
         title: 'New Project',
@@ -33,7 +30,7 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        home: RootScreen(),
+        home: RootScreenT(),
       );
     });
   }
