+++ lib/ui/screens/contentPlayer/content_player_screen.dart
@@ -1,7 +1,7 @@
 /* 해당 소스파일에서 MultiImport 방식으로 Import 구문 작성 필요, YoutubePlayer, YoutubePlayerIframe 패키지가 중복되어 에러 발생 */
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:movie_curation/widgets/back_arrow_button.dart';
+import 'package:movie_curation/ui/widgets/back_arrow_button.dart';
 import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
 class ContentPlayerScreen extends StatelessWidget {
