+++ lib/ui/screens/movieDetail/localWidget/main_content_info_mobile.dart
