+++ lib/ui/screens/movieDetail/movie_detail_mobile.dart
