+++ lib/ui/screens/movieDetail/movie_detail_tablet.dart
