+++ lib/ui/screens/searchScreen/localWidget/content_thumbnail_vertical_slider.dart
