+++ lib/ui/screens/searchScreen/search_screen_mobile.dart
