+++ lib/ui/widgets/movie_content_info.dart
