+++ lib/ui/widgets/movie_post_slider.dart
