+++ lib/ui/widgets/trailer_dialog.dart
@@ -1,4 +1,4 @@
-import 'package:movie_curation/screens/home/mobile/trailer_player_screen_mobile.dart';
+import 'package:movie_curation/ui/screens/home/mobile/trailer_player_screen_mobile.dart';
 import 'package:movie_curation/utilities/index.dart';
 
 // // TODO: Iframe 추가 설정 필요(소리, 테마 컬러 등) + 반응형 Margin
