+++ lib/utilities/index.dart
@@ -15,19 +15,17 @@ export 'package:clickable_list_wheel_view/clickable_list_wheel_widget.dart';
 export 'package:carousel_slider/carousel_slider.dart';
 
 /* Screens */
-export 'package:movie_curation/screens/root/root_screen.dart';
-export 'package:movie_curation/screens/home/tablet/home_page_builder_tablet.dart';
-export 'package:movie_curation/screens/home/mobile/home_screen_mobile.dart';
-export 'package:movie_curation/screens/home/tablet/home_screen_tablet.dart';
+export 'package:movie_curation/ui/screens/home/tablet/home_page_builder_tablet.dart';
+export 'package:movie_curation/ui/screens/home/mobile/home_screen_mobile.dart';
+export 'package:movie_curation/ui/screens/home/tablet/home_screen_tablet.dart';
 export 'package:movie_curation/utilities/temp_screen.dart';
 export 'package:movie_curation/utilities/temp_screen.dart';
-export 'package:movie_curation/screens/movieDetail/movie_detail_tablet.dart';
-export 'package:movie_curation/screens/movieDetail/movie_detail_mobile.dart';
+export 'package:movie_curation/ui/screens/movieDetail/movie_detail_tablet.dart';
+export 'package:movie_curation/ui/screens/movieDetail/movie_detail_mobile.dart';
 export 'package:youtube_player_flutter/youtube_player_flutter.dart';
-export 'package:movie_curation/screens/contentPlayer/content_player_screen.dart';
-export 'package:movie_curation/screens/root/root_screen_mobile.dart';
-export 'package:movie_curation/screens/searchScreen/search_screen_tablet.dart';
-export 'package:movie_curation/screens/searchScreen/search_screen_mobile.dart';
+export 'package:movie_curation/ui/screens/contentPlayer/content_player_screen.dart';
+export 'package:movie_curation/ui/screens/searchScreen/search_screen_tablet.dart';
+export 'package:movie_curation/ui/screens/searchScreen/search_screen_mobile.dart';
 
 /* Models */
 export 'package:movie_curation/models/movie_model.dart';
@@ -47,21 +45,21 @@ export 'package:movie_curation/viewModels/youtube_view_model.dart';
 export 'package:movie_curation/viewModels/search_content_view_model.dart';
 
 /* Widgets */
-export 'package:movie_curation/widgets/responsive_layout.dart';
-export 'package:movie_curation/widgets/gradient_button.dart';
-export 'package:movie_curation/widgets/category_group_button.dart';
-export 'package:movie_curation/widgets/gradient_post_background.dart';
-export 'package:movie_curation/widgets/movie_content_info.dart';
-export 'package:movie_curation/widgets/movie_post_slider.dart';
-export 'package:movie_curation/widgets/trailer_dialog.dart';
-export 'package:movie_curation/screens/movieDetail/localWidget/cast_slider.dart';
-export 'package:movie_curation/screens/movieDetail/localWidget/movie_else_info.dart';
-export 'package:movie_curation/widgets/back_arrow_button.dart';
-export 'package:movie_curation/screens/movieDetail/localWidget/main_content_info_mobile.dart'; // MovieDetailScreen(M) > MainContentInfoMobile
-export 'package:movie_curation/screens/movieDetail/localWidget/youtube_review_contents_wheel_slider.dart'; // MovieDetailScreen(T) > WheelSlider
-export 'package:movie_curation/screens/searchScreen/localWidget/random_poster_background.dart';
-export 'package:movie_curation/screens/searchScreen/localWidget/content_thumbnail_vertical_slider.dart';
-export 'package:movie_curation/screens/movieDetail/localWidget/youtube_content_list_view.dart';
+export 'package:movie_curation/ui/widgets/responsive_layout.dart';
+export 'package:movie_curation/ui/widgets/gradient_button.dart';
+export 'package:movie_curation/ui/widgets/category_group_button.dart';
+export 'package:movie_curation/ui/widgets/gradient_post_background.dart';
+export 'package:movie_curation/ui/widgets/movie_content_info.dart';
+export 'package:movie_curation/ui/widgets/movie_post_slider.dart';
+export 'package:movie_curation/ui/widgets/trailer_dialog.dart';
+export 'package:movie_curation/ui/screens/movieDetail/localWidget/cast_slider.dart';
+export 'package:movie_curation/ui/screens/movieDetail/localWidget/movie_else_info.dart';
+export 'package:movie_curation/ui/widgets/back_arrow_button.dart';
+export 'package:movie_curation/ui/screens/movieDetail/localWidget/main_content_info_mobile.dart'; // MovieDetailScreen(M) > MainContentInfoMobile
+export 'package:movie_curation/ui/screens/movieDetail/localWidget/youtube_review_contents_wheel_slider.dart'; // MovieDetailScreen(T) > WheelSlider
+export 'package:movie_curation/ui/screens/searchScreen/localWidget/random_poster_background.dart';
+export 'package:movie_curation/ui/screens/searchScreen/localWidget/content_thumbnail_vertical_slider.dart';
+export 'package:movie_curation/ui/screens/movieDetail/localWidget/youtube_content_list_view.dart';
 
 /* Services */
 export 'package:movie_curation/services/movie_api_response.dart';
